shorts_pipeline/requirements.txt
pytrends
requests
beautifulsoup4
openai>=1.30
python-dotenv
pandas
